# legalai-backend/app/static/openapi.yaml
openapi: 3.0.3
info:
  title: Legal Lawyer AI API
  version: "1.0.0"
  description: Swagger spec generated from the provided Postman collection.
servers:
  - url: http://localhost:5000
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    AnyObject:
      type: object
      additionalProperties: true
    Error:
      type: object
      properties:
        message: { type: string }
        error: { type: string }
        details: { $ref: "#/components/schemas/AnyObject" }
...
security: []  # default none; per-route add as needed
paths:
...
  /api/v1/auth/signup:
    post:
      tags: [Auth]
      summary: Signup
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AnyObject" }
            example:
              name: Ahmed Khan
              email: ahmed@example.com
              phone: "03001234567"
              cnic: "12345-1234567-1"
              password: "SecurePass@123"
              city: Lahore
              gender: male
              age: 28
              fatherName: Khan Sahib
              fatherCnic: "11111-1111111-1"
              motherName: Begum Sahiba
              motherCnic: "22222-2222222-2"
              totalSiblings: 3
              brothers: 2
              sisters: 1
              timezone: Asia/Karachi
      responses:
        "200": { description: Success, content: { application/json: { schema: { $ref: "#/components/schemas/AnyObject" } } } }
        "400": { description: Bad Request, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
  /api/v1/auth/login:
    post:
      tags: [Auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AnyObject" }
            example:
              email: ahmed@example.com
              password: "SecurePass@123"
      responses:
        "200": { description: Success, content: { application/json: { schema: { $ref: "#/components/schemas/AnyObject" } } } }
        "401": { description: Unauthorized, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
  /api/v1/auth/refresh:
    post:
      tags: [Auth]
      summary: Refresh token
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AnyObject" }
            example:
              refreshToken: YOUR_REFRESH_TOKEN_HERE
      responses:
        "200": { description: Success, content: { application/json: { schema: { $ref: "#/components/schemas/AnyObject" } } } }
        "401": { description: Unauthorized, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
  /api/v1/auth/verify-email:
    get:
      tags: [Auth]
      summary: Verify email
      parameters:
        - in: query
          name: token
          required: true
          schema: { type: string }
      responses:
        "200": { description: Success, content: { application/json: { schema: { $ref: "#/components/schemas/AnyObject" } } } }
        "400": { description: Invalid token, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
  /api/v1/auth/forgot-password:
    post:
      tags: [Auth]
      summary: Forgot password
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AnyObject" }
            example:
              email: ahmed@example.com
      responses:
        "200": { description: Success, content: { application/json: { schema: { $ref: "#/components/schemas/AnyObject" } } } }
  /api/v1/auth/reset-password:
    post:
      tags: [Auth]
      summary: Reset password
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AnyObject" }
            example:
              token: YOUR_RESET_TOKEN_HERE
              newPassword: "NewSecure@Pass456"
      responses:
        "200": { description: Success, content: { application/json: { schema: { $ref: "#/components/schemas/AnyObject" } } } }
  /api/v1/auth/change-password:
    post:
      tags: [Auth]
      summary: Change password
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AnyObject" }
            example:
              currentPassword: "SecurePass@123"
              newPassword: "NewSecure@Pass789"
              confirmPassword: "NewSecure@Pass789"
      responses:
        "200": { description: Success, content: { application/json: { schema: { $ref: "#/components/schemas/AnyObject" } } } }
        "401": { description: Unauthorized, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }

  # -------------------- USERS --------------------
...
  /api/v1/chat/ask:
    post:
      tags: [Chat]
      summary: Ask question (new or existing conversation)
      security: [{ bearerAuth: [] }]
      parameters:
        - in: header
          name: X-Safe-Mode
          required: false
          schema: { type: string }
          description: Set to "1" to enable safe mode.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AnyObject" }
            examples:
              newConversation:
                value:
                  question: "What are the divorce laws in Pakistan?"
                  language: en
              existingConversation:
                value:
                  question: "What documents are required?"
                  language: en
                  conversationId: 1
      responses:
        "200": { description: Success }
  /api/v1/chat/conversations:
    get:
      tags: [Chat]
      summary: List conversations
      security: [{ bearerAuth: [] }]
      parameters:
        - in: query
          name: page
          schema: { type: integer, default: 1 }
        - in: query
          name: limit
          schema: { type: integer, default: 20 }
      responses:
        "200": { description: Success }
  /api/v1/chat/conversations/{conversationId}/messages:
    get:
      tags: [Chat]
      summary: Get conversation messages
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: conversationId
          required: true
          schema: { type: integer }
        - in: query
          name: page
          schema: { type: integer, default: 1 }
        - in: query
          name: limit
          schema: { type: integer, default: 30 }
      responses:
        "200": { description: Success }
  /api/v1/chat/conversations/{conversationId}:
    put:
      tags: [Chat]
      summary: Rename conversation
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: conversationId
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AnyObject" }
            example:
              title: "Divorce Law Discussion"
      responses:
        "200": { description: Success }
    delete:
      tags: [Chat]
      summary: Delete conversation
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: conversationId
          required: true
          schema: { type: integer }
      responses:
        "200": { description: Success }

  # -------------------- REMINDERS --------------------
...